+++ src/main.rs
@@ -4,7 +4,7 @@ use wden::profile::ProfileStore;
 fn try_read_profile_name(input: &str) -> Result<String, anyhow::Error> {
     if input
         .chars()
-        .any(|c| !c.is_ascii_lowercase() && !c.is_digit(10) && c != '_' && c != '-')
+        .any(|c| !c.is_ascii_lowercase() && !c.is_ascii_digit() && c != '_' && c != '-')
     {
         Err(anyhow::anyhow!("Invalid profile name. Profile names can only include lowercase alphanumeric characters, dashes (-) and underscores (_)."))
     } else {
@@ -32,6 +32,14 @@ struct Opts {
     /// list all stored profiles
     #[clap(long)]
     list_profiles: bool,
+
+    /// Accept invalid and untrusted (e.g. self-signed) certificates
+    /// when connecting to the server. This option makes connections
+    /// insecure, so avoid using it. Note: this option is not stored
+    /// in the profile settings. It must be specified every time when
+    /// connecting to servers with untrusted certificates.
+    #[clap(long)]
+    accept_invalid_certs: bool,
 }
 
 #[tokio::main]
@@ -43,7 +51,7 @@ async fn main() {
         return;
     }
 
-    wden::ui::launch(opts.profile, opts.server_url);
+    wden::ui::launch(opts.profile, opts.server_url, opts.accept_invalid_certs);
 }
 
 fn list_profiles() -> std::io::Result<()> {