+++ src/ui/item_details.rs
@@ -10,7 +10,7 @@ use crate::{
     ui::components::secret_text_view::SecretTextView,
 };
 use cursive::{
-    theme::{BaseColor, Color, Effect, Style, ColorStyle},
+    theme::{BaseColor, Color, ColorStyle, Effect, Style},
     traits::{Nameable, Resizable},
     view::Margins,
     views::{Dialog, LinearLayout, OnEventView, PaddedView, ScrollView, TextView, ViewRef},
@@ -20,8 +20,7 @@ use lazy_static::lazy_static;
 use log::warn;
 
 lazy_static! {
-    static ref VALUE_STYLE: Style =
-        Style::from(Effect::Reverse).combine(ColorStyle::secondary());
+    static ref VALUE_STYLE: Style = Style::from(Effect::Reverse).combine(ColorStyle::secondary());
 }
 
 pub fn item_detail_dialog(ud: &StatefulUserData<Unlocked>, item_id: &str) -> Option<impl View> {