+++ src/ui/launch.rs
@@ -1,13 +1,16 @@
 use std::sync::Arc;
 
-use cursive::{Cursive, CursiveRunnable, theme::Theme, theme::PaletteColor::*, theme::BaseColor, theme::Color};
+use cursive::{
+    theme::BaseColor, theme::Color, theme::PaletteColor::*, theme::Theme, Cursive, CursiveRunnable,
+};
 
 use crate::profile::{GlobalSettings, ProfileData, ProfileStore};
 
 use super::{autolock, data::UserData, login::login_dialog};
 
-pub fn launch(profile: String, server_url: Option<String>) {
-    let (global_settings, profile_data, profile_store) = load_profile(profile, server_url);
+pub fn launch(profile: String, server_url: Option<String>, accept_invalid_certs: bool) {
+    let (global_settings, profile_data, profile_store) =
+        load_profile(profile, server_url, accept_invalid_certs);
     let profile_name = global_settings.profile.clone();
 
     let mut siv = cursive::default();
@@ -24,7 +27,7 @@ pub fn launch(profile: String, server_url: Option<String>) {
     cursive::logger::init();
     log::set_max_level(log::LevelFilter::Info);
 
-    siv.add_layer(login_dialog(&profile_name, profile_data.saved_email));
+    siv.add_layer(login_dialog(&profile_name, profile_data.saved_email, false));
 
     run(siv);
 }
@@ -53,6 +56,7 @@ fn run(mut cursive: CursiveRunnable) {
 fn load_profile(
     profile_name: String,
     server_url: Option<String>,
+    accept_invalid_certs: bool,
 ) -> (GlobalSettings, ProfileData, ProfileStore) {
     let profile_store = ProfileStore::new(&profile_name);
     let mut profile_data = profile_store.load().unwrap_or_default();
@@ -62,6 +66,7 @@ fn load_profile(
         server_url: server_url.unwrap_or(profile_data.server_url),
         autolock_duration: profile_data.autolock_duration,
         device_id: profile_data.device_id.clone(),
+        accept_invalid_certs,
     };
 
     // Write new settings